dom/html/test/test_imageSrcSet.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=980243
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 980243</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 980243 **/
  SimpleTest.waitForExplicitFinish();

  addLoadEvent(function() {
    var img = document.querySelector("img");
    img.onload = function() {
      ok(true, "Reached here");
      SimpleTest.finish();
    }
    // If ths spec ever changes to treat .src sets differently from
    // setAttribute("src"), we'll need some sort of canonicalization step
    // earlier to make the attr value an absolute URI.
    img.setAttribute("src", img.getAttribute("src"));
  });
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=980243">Mozilla Bug 980243</a>
<p id="display"></p>
<div id="content" style="display: none">
  <img src="file_formSubmission_img.jpg">
</div>
<pre id="test">
</pre>
</body>
</html>
